pyproject.toml
[project]
name = "agentic-web"
version = "0.2.0"
description = "Multi-Agent A2A System with LLM Integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.11",
    "uvicorn>=0.30.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "aiofiles>=23.0.0",
]
